@@ -4,7 +4,7 @@
     "license": "MIT",
     "description": "php SDK to provide a friendly interface to the payFURL API",
     "require": {
-        "php": ">=7.3",
-        "another-vendor/package": "1.*"
+        "php": ">=7.4",
+        "ext-curl": "*"
     }
 }
