@@ -1,4 +1,5 @@
 <?php
+
 namespace payFURL\Sdk;
 
 use payFURL\Sdk\Config;
@@ -7,79 +8,78 @@ use payFURL\Sdk\ResponseException;
 require_once(__DIR__ . "/../Config.php");
 require_once(__DIR__ . "/../ResponseException.php");
 
-/*
- * (c) payFURL
+/**
+ * @copyright PayFURL
  */
+
 class HttpWrapper
 {
-    static function CallApi($Endpoint, $Method, $Body)
+    static function CallApi($endpoint, $method, $body, $addHeaders = [])
     {
-        $Url = Config::$BaseUrl . $Endpoint;
+        $url = Config::$BaseUrl . $endpoint;
 
-        if (Config::$EnableDebug)
-        {
-            print "calling URL: " . $Url . "\n";
+        if (Config::$EnableDebug) {
+            print "calling URL: " . $url . "\n";
             print "Timeout: " . Config::$TimeoutMilliseconds . "\n";
         }
 
         $ch = curl_init();
-        curl_setopt($ch, CURLOPT_CUSTOMREQUEST, $Method);
-        curl_setopt($ch, CURLOPT_POSTFIELDS, $Body);
-        curl_setopt($ch, CURLOPT_URL, $Url);
+        curl_setopt($ch, CURLOPT_CUSTOMREQUEST, $method);
+        curl_setopt($ch, CURLOPT_POSTFIELDS, $body);
+        curl_setopt($ch, CURLOPT_URL, $url);
         curl_setopt($ch, CURLOPT_ENCODING, "gzip");
         curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
         curl_setopt($ch, CURLOPT_SSLVERSION, CURL_SSLVERSION_TLSv1_2);
         curl_setopt($ch, CURLOPT_TIMEOUT_MS, Config::$TimeoutMilliseconds);
 
-        if (strtolower(Config::$Environment) == "local")
-        {
+        if (strtolower(Config::$Environment) == "local") {
             curl_setopt($ch, CURLOPT_SSL_VERIFYHOST, false);
             curl_setopt($ch, CURLOPT_SSL_VERIFYPEER, false);
         }
 
-        $Headers = [
+        $headers = [
             "Content-Type: application/json",
-            "Content-Length: " . strlen($Body),
-            "x-secretkey:" . Config::$SecretKey
+            "Content-Length: " . strlen($body),
+            "x-secretkey:" . Config::$SecretKey,
         ];
+        if (count($addHeaders) > 0) {
+            foreach ($addHeaders as $key => $value) {
+                $headers[] = $key . ':' . $value;
+            }
+        }
 
-        curl_setopt($ch, CURLOPT_HTTPHEADER, $Headers);
+        curl_setopt($ch, CURLOPT_HTTPHEADER, $headers);
 
-        $Response = curl_exec($ch);
-        $Info = curl_getinfo($ch);
-        $Error = curl_errno($ch);
+        $response = curl_exec($ch);
+        $info = curl_getinfo($ch);
+        $error = curl_errno($ch);
         curl_close($ch);
 
         // hande timeout
-        if ($Info["http_code"] == 0)
-        {
-            if (Config::$EnableDebug)
-            {
-                var_dump($Response);
-                var_dump($Info);
-                var_dump($Error);
+        if ($info["http_code"] == 0) {
+            if (Config::$EnableDebug) {
+                var_dump($response);
+                var_dump($info);
+                var_dump($error);
             }
             throw new ResponseException("Request Timeout", 408);
         }
 
         // error handling
-        if ($Info["http_code"] != 200 && $Info["http_code"] != 201)
-        {
-            $ResponseJson = json_decode($Response, true);
-            $Message = "";
-            if (array_key_exists("message", $ResponseJson))
-            {
-                $Message = $ResponseJson["message"];
+        if ($info["http_code"] != 200 && $info["http_code"] != 201) {
+            $responseJson = json_decode($response, true);
+            $message = "";
+            if (array_key_exists("message", $responseJson)) {
+                $message = $responseJson["message"];
             }
-            if (Config::$EnableDebug)
-            {
-                var_dump($Response);
-                var_dump($Info);
-                var_dump($Error);
+            if (Config::$EnableDebug) {
+                var_dump($response);
+                var_dump($info);
+                var_dump($error);
             }
-            throw new ResponseException($Message, $Info["http_code"]);
+            throw new ResponseException($message, $info["http_code"]);
         }
 
-        return json_decode($Response, true);
+        return json_decode($response, true);
     }
 }
