@@ -1,10 +1,11 @@
 <?php declare(strict_types=1);
+
 use PHPUnit\Framework\TestCase;
 
-require_once(__DIR__ . "/../src/Config.php");
-require_once(__DIR__ . "/../src/Charge.php");
-require_once(__DIR__ . "/TestBase.php");
-require_once(__DIR__ . "/../src/ResponseException.php");
+require_once(__DIR__ . '/../src/Config.php');
+require_once(__DIR__ . '/../src/Charge.php');
+require_once(__DIR__ . '/TestBase.php');
+require_once(__DIR__ . '/../src/ResponseException.php');
 
 use payFURL\Sdk\Config;
 use payFURL\Sdk\Charge;
@@ -12,39 +13,51 @@ use payFURL\Sdk\ResponseException;
 
 final class ChargeTest extends TestBase
 {
+    /**
+     * @throws ResponseException
+     */
     public function testCreateWithCard(): void
     {
         $svc = new Charge();
 
         $result = $svc->CreateWithCard([
-            "Amount" => 15.5,
-            "Currency" => "AUD",
-            "Reference" => "123",
-            "ProviderId" => $this->CardProviderId,
-            "CardNumber" => "4111111111111111",
-            "ExpiryDate" => "10/30",
-            "Ccv" => "123",
-            "Cardholder" => "Test Cardholder"]);
+                                           'Amount' => 15.5,
+                                           'Currency' => 'AUD',
+                                           'Reference' => '123',
+                                           'ProviderId' => $this->CardProviderId,
+                                           'PaymentInformation' => [
+                                               'CardNumber' => '4111111111111111',
+                                               'ExpiryDate' => '10/30',
+                                               'Ccv' => '123',
+                                               'Cardholder' => 'Test Cardholder'
+                                           ]]);
 
-        $this->assertSame('SUCCESS', $result["status"]);
+        $this->assertSame('SUCCESS', $result['status']);
     }
 
+    /**
+     * @throws ResponseException
+     */
     public function testCreateWithCardLeastCost(): void
     {
         $svc = new Charge();
 
         $result = $svc->CreateWithCardLeastCost([
-            "Amount" => 15.5,
-            "Currency" => "AUD",
-            "Reference" => "123",
-            "CardNumber" => "4111111111111111",
-            "ExpiryDate" => "10/30",
-            "Ccv" => "123",
-            "Cardholder" => "Test Cardholder"]);
+                                                    'Amount' => 15.5,
+                                                    'Currency' => 'AUD',
+                                                    'Reference' => '123',
+                                                    'PaymentInformation' => [
+                                                        'CardNumber' => '4111111111111111',
+                                                        'ExpiryDate' => '10/30',
+                                                        'Ccv' => '123',
+                                                        'Cardholder' => 'Test Cardholder']]);
 
-        $this->assertSame('SUCCESS', $result["status"]);
+        $this->assertSame('SUCCESS', $result['status']);
     }
 
+    /**
+     * @throws ResponseException
+     */
     public function testWithInvalidProvider(): void
     {
         $svc = new Charge();
@@ -52,16 +65,20 @@ final class ChargeTest extends TestBase
         $this->expectException(ResponseException::class);
 
         $result = $svc->CreateWithCard([
-            "Amount" => 15.5,
-            "Currency" => "AUD",
-            "Reference" => "123",
-            "ProviderId" => "invalid_provider",
-            "CardNumber" => "4111111111111111",
-            "ExpiryDate" => "10/30",
-            "Ccv" => "123",
-            "Cardholder" => "Test Cardholder"]);
+                                           'Amount' => 15.5,
+                                           'Currency' => 'AUD',
+                                           'Reference' => '123',
+                                           'ProviderId' => 'invalid_provider',
+                                           'PaymentInformation' => [
+                                               'CardNumber' => '4111111111111111',
+                                               'ExpiryDate' => '10/30',
+                                               'Ccv' => '123',
+                                               'Cardholder' => 'Test Cardholder']]);
     }
 
+    /**
+     * @throws ResponseException
+     */
     public function testWithShortTimeout(): void
     {
         $svc = new Charge();
@@ -73,79 +90,96 @@ final class ChargeTest extends TestBase
         Config::$TimeoutMilliseconds = $Timeout;
 
         $result = $svc->CreateWithCard([
-            "Amount" => 15.5,
-            "Currency" => "AUD",
-            "Reference" => "123",
-            "ProviderId" => $this->CardProviderId,
-            "CardNumber" => "4111111111111111",
-            "ExpiryDate" => "10/30",
-            "Ccv" => "123",
-            "Cardholder" => "Test Cardholder"]);
+                                           'Amount' => 15.5,
+                                           'Currency' => 'AUD',
+                                           'Reference' => '123',
+                                           'ProviderId' => $this->CardProviderId,
+                                           'PaymentInformation' => [
+                                               'CardNumber' => '4111111111111111',
+                                               'ExpiryDate' => '10/30',
+                                               'Ccv' => '123',
+                                               'Cardholder' => 'Test Cardholder']]);
     }
 
+    /**
+     * @throws ResponseException
+     */
     public function testSingle(): void
     {
         $svc = new Charge();
 
         $chargeResult = $svc->CreateWithCard([
-            "Amount" => 15.5,
-            "Currency" => "AUD",
-            "Reference" => "123",
-            "ProviderId" => $this->CardProviderId,
-            "CardNumber" => "4111111111111111",
-            "ExpiryDate" => "10/30",
-            "Ccv" => "123",
-            "Cardholder" => "Test Cardholder"]);
+                                                 'Amount' => 15.5,
+                                                 'Currency' => 'AUD',
+                                                 'Reference' => '123',
+                                                 'ProviderId' => $this->CardProviderId,
+                                                 'PaymentInformation' => [
+                                                     'CardNumber' => '4111111111111111',
+                                                     'ExpiryDate' => '10/30',
+                                                     'Ccv' => '123',
+                                                     'Cardholder' => 'Test Cardholder']]);
 
-        $singleResult = $svc->Single(["ChargeId" => $chargeResult["chargeId"]]);
-        
-        $this->assertSame($chargeResult["chargeId"], $singleResult["chargeId"]);
+        $singleResult = $svc->Single(['ChargeId' => $chargeResult['chargeId']]);
+
+        $this->assertSame($chargeResult['chargeId'], $singleResult['chargeId']);
     }
 
+    /**
+     * @throws ResponseException
+     */
     public function testRefund(): void
     {
         $svc = new Charge();
 
         $chargeResult = $svc->CreateWithCard([
-            "Amount" => 15.5,
-            "Currency" => "AUD",
-            "Reference" => "123",
-            "ProviderId" => $this->CardProviderId,
-            "CardNumber" => "4111111111111111",
-            "ExpiryDate" => "10/30",
-            "Ccv" => "123",
-            "Cardholder" => "Test Cardholder"]);
+                                                 'Amount' => 15.5,
+                                                 'Currency' => 'AUD',
+                                                 'Reference' => '123',
+                                                 'ProviderId' => $this->CardProviderId,
+                                                 'PaymentInformation' => [
+                                                     'CardNumber' => '4111111111111111',
+                                                     'ExpiryDate' => '10/30',
+                                                     'Ccv' => '123',
+                                                     'Cardholder' => 'Test Cardholder']]);
 
-        $refundResult = $svc->Refund(["ChargeId" => $chargeResult["chargeId"], "Amount" => 5]);
-        
-        $this->assertSame(5, $refundResult["refundedAmount"]);
+        $refundResult = $svc->Refund(['ChargeId' => $chargeResult['chargeId'], 'Amount' => 5]);
+
+        $this->assertSame(5, $refundResult['refundedAmount']);
     }
 
+    /**
+     * @throws ResponseException
+     * @throws Exception
+     */
     public function testSearch(): void
     {
         $svc = new Charge();
 
         $Reference = bin2hex(random_bytes(16));
         $result = $svc->CreateWithCard([
-            "Amount" => 15.5,
-            "Currency" => "AUD",
-            "Reference" => $Reference,
-            "ProviderId" => $this->CardProviderId,
-            "CardNumber" => "4111111111111111",
-            "ExpiryDate" => "10/30",
-            "Ccv" => "123",
-            "Cardholder" => "Test Cardholder"]);
+                                           'Amount' => 15.5,
+                                           'Currency' => 'AUD',
+                                           'Reference' => $Reference,
+                                           'ProviderId' => $this->CardProviderId,
+                                           'PaymentInformation' => [
+                                               'CardNumber' => '4111111111111111',
+                                               'ExpiryDate' => '10/30',
+                                               'Ccv' => '123',
+                                               'Cardholder' => 'Test Cardholder']]);
 
-        $searchResult = $svc->Search(array("Reference" => $Reference));
-        
-        $this->assertSame(1, $searchResult["count"]);
+        $searchResult = $svc->Search(array('Reference' => $Reference));
+
+        $this->assertSame(1, $searchResult['count']);
     }
 
+    /**
+     * @throws ResponseException
+     */
     public function testInvalidParameters(): void
     {
         $svc = new Charge();
         $this->expectException(ResponseException::class);
 
-        $result = $svc->CreateWithCard(["Amount" => 15.5]);
+        $result = $svc->CreateWithCard(['Amount' => 15.5]);
     }
 }
